requirements.txt
django==1.4.2
django-cms==2.3.3
south==0.7.6
MySQL-python==1.2.3
Pillow==1.7.5

django-blog-zinnia==0.11.2

# CMS Nice to Have
django-reversion==1.6.3
django-sekizai==0.6.1

# CMS Plugins We'll Want:
cmsplugin-contact==1.0.0
cmsplugin-zinnia==0.3
#cmsplugin-news==0.3c

# Newsletter
# django-newsletter==0.2.6
# django-ses==0.4.
# django-tinymce==1.5.1b2
